imaging: send the PXE inventory to the inventory-server
The PXE inventory wasn't written to the inventory at all, except the
MACAdress.

This inventory is now send to the inventory-server configured in the
imaging-server configuration like an OCS inventory so that it can be
compatible with OCS servers.

We can also benefit of the entity/client mapping provided by the
inventory-server with no OCS agent installed.

If the inventory can't be sent for some reason (inventory-server down)
this won't stop the client to be up-to-date in the imaging database.