scheduler : isolate the proxy of incomming launcher requests
When the scheduler is on huge load (e.g. big SQL queries,
sending many requests to launcher, etc), the proxy processing
incomming requests from launcher has no more the ressources
to working - incomming requests are timeouted.

This proxy is now executed in own process when the main run-time
of scheduler is started.

+ For the same reasons, the preempting of commands is executed
in own thread.