Fixes #1989: If installation failed, install files left on client
Added a shortcut on workflow when a command on host fails on execution
during all attempts. This task will be rescheduled with default command
life-time and number of attempts, having delete phase only.

When this phase is finished, this task will be terminated allways as failed.
(deletion phase still logged as standard step, visible on history tab)