glpi: Fix machines are missing in computers list
total machine count can be different in computer list or in inventory
dashboard.

In computers list, a filt['hostname'] is passed and if it's empty, it
add a filter: glpi_computers.name LIKE '%'

With this filter, not all machines could not be displayed, so if
filt['hostname'] is empty, don't add this filter on query.