glpi: manufacturer warranty pages extended support
Some manufacturer pages requires the serial number to be posted. This commit
allows us to do so if needed. Default is still a simple url link.

Use type = post in the glpi.ini configuration to generate a form to be posted
to the manufacturer warrranty page.

Unfortunately posting data to the HP warranty page doesn't work, but works for
Lenovo.

Closes #2220