(sql) Fix schema63.sql to be replayed.
The schema 63 wasn't idempotent.
As the root login is uniq we cannot create is more than once. We ignore
the error if it already exists.
The ind_date_log index is only created if it does not already exists.
Add a missing semi-colon when creating the support_help_command table.