(deploy) Fix use of parsexmppjsonfile function.
The parsexmppjsonfile is used to change False/True to false/true in
xmppdeploy.json

Before we were doing it while deploying, but as now we can have several
deployement subsitutes, we got concurrent access to the file and it resulted in empty
xmppdeploy.json.

Now we only use this function while saving the xmppdeploy.json file.