(glpi 9.5) Fix crash with some incomplete inventories
It can happen that some versions ( only seen with ubuntu for the moment
) are missing in Glpi.

In consequence machine['version'] returns None.

I added a workaroung, and now if machine['version'] is equal to None, we
give a fake version 00.00.

It allow to fix the crash and should make simpler to detect uncomplete
inventories