(sql) Fix sessionname size in the historylog table.
In the commit 0d9ad0b88af82ffb8ec86b9c0cd09ae772704bb1 we changed the
size of the sessionname entry in the logs table to be a varchar(45).

This commit make it coherent with the historylog table as the historylog
is a copy of the log table.
This way we make sure we do not have any error nor data loss.