(ui) Add a Search button on every pages.
When we start to enter text in the search bar, when we reach 3
characters, it start to query the database. It costs on the server.

With the new behaviour it won't query the db until we click on the
button

(cherry picked from commit dfa4be0516d9177e74b234a99caa5ffac90debf7)