Fixes to Kiosk feature (#368)
* Remove useless whitespaces

* Launches the appropriate section of the selected package

The install button deploy the install section of the package;
The update button deploy the update section of the package;
The delete button deploy the uninstall section of the package.

* Get the sent date for install / uninstall

When we install or uninstall a package the plugin get the start datetime from the kiosk.

* Add launch section into package interface

* fix missing import

* add missing function

the function utc2local is missing from xmppmaste/lib/utils

* Add comment and imports

The utc2local function was not commented and used some missing functions.

* Add a launcher input for the packages

This input specifies the app launcher needed by the kiosk.

* The base64 encoding was wrong

The launcher string is encoded in base64. But for some issues this one was not correctly encoded / decoded.

* Add the package uuid to the descriptor

The uuid is added to the metaparameter of the package descriptor.

* The old_launcher create traceback

When a form is sent to POST request, all the datas are copied in the old_data.
If the element contained in the data is encoded because of special chars, the old_data is not.
This generates a traceback.

* Send message when a launcher is updated

When a package is updated, a message which contains the

* Add action status to package

This action is usefull to send the avancment stat when a deployment is launched.
The status into install, update or uninstall section are independants

* Update the changelog

Update the changelog for the kiosk.

* Do not try to install pulse2-imaging-server

* Revert "Add a deps on pulse-imaging-scripts"

This reverts commit a8c8bf3d6a26b61997bcae6297cd5dbc9fdc1343.

* Modify the structure of data message

The structure of the data key into the message sent to the Agent Machine, is complexified.
Now An action is systematically specified

* fix wrong variable name

* Update the kiosk presence

If the kiosk is detected on the machine, the presence is set to "True". Else the presence is set to "False"

* Remove debug print

This line generate a big amount of data into the prompt

* Add a notification action

This action sent notification message and / or deployment info to the User or the Machine or the Kiosk

* Preselect the type of notification

In this action, the type of the notification were not automatically preselected.

* The installations are launched at the specified date

The installations were not launched at the specified date. This problem is fixed

* Add missing imports

The time and datetime libs are missing.

* remove Associate packages from profiles list

This action is declared but unused. So It is removed

* Update the notification action

* Add the package name to the log

The section (install/update/delete) is specified with the name of the package.

* Add recursive_paths method.

This method get recursively the children paths   of the tree.

* Add a function which returns the general tree of all OUs

This function returns the tree object of all the OUs found.

* Search the profile related to the user's parent OU.

* Add missing import

* Add a function to notify a single machine.

This function send the profiles to the specified machine when it is invoked.

* Update kiosk packages list

When an inventory is comming from a machine AND this machine has the kiosk loaded, the packages list for the kiosk is updated.

* Add disply_name parameter for each element in the topology

This element is used for display. The historical name is still used for the links.

* [kiosk] Use specific suffix_ou parameter for OUs

* [kiosk] Enable kiosk by default

Co-authored-by: botheis <noel.kneib@siveo.net>
Co-authored-by: Nicolas Lécureuil <neoclust@mageia.org>