Fix generation of hashes when a file has been deleted from the package (#441)
* (cdn) Fix hash generation in case of remove file in package

Before, when we remove a file in package, the hash was not regenerated, so it's make a hash invalid error.
Now, if we remove or edit file in package, re genere hash and remove hash invalid error.

* (cdn) Fix concat path and extension file

In this case we use os.path.join, it will put a "/" between dest path and .hash extension file, so we just concat dest path and .hash.

---------

Co-authored-by: William Vasselin <43633092+Hruund@users.noreply.github.com>