(imaging) Fix support for new clonezilla for multicast
Here we had a double issue we needed to add ipxe support and fix
multicast.

The main issue we found is that clonezilla does not yet support iPXE.
We decided to create a dumb default file using old PXELinux syntax and a
file with iPXE syntax for the real boot.

As clonezilla want its file to be called default, we changed the name of
our iPXE file to medulla.

TODO: Modify in ipxe-siveo package too.