(deployment) Do not allow wrongly defined intervals (#485)
When the user define an interterval we check if it is correct.
If not we show a popup and deactive the submit button, for exemple if
the first element is greater than the second

I.E.:
interval 19-5 will disable the button
and on intervals [a-b] where a > b, change to [a-24,0-b] intervals.

Co-authored-by: botheis <noel.kneib@siveo.net>