(FEATURE) Use iPXE new way of using menus.
Previously we used text files to send menus to iPXE ( the same way we
did with PXELinux ).

Now, with iPXE we can store all in the SQL Database and retrieve it
though SQL Request via a PHP Webpage.

It will remove bugs related to files synchronisation between relays.

If the KEYBOARD_MAP is set to dynamic, the keyboard will be handled.
Source : https://ipxe.org/buildcfg/keyboard_map