(urbackup) fix first initialization computer on urbackup
On this first initialization, we will check if user already exist on urbackup and mysql DB and if he is in urbackup profile.

If computer already exist but have not profile, we only search his internet auth key with API and insert it on mysql DB and assign it a choosen profile.
If computer not exist on any DB, we will use API to insert it on urbackup DB and assign it a choosen profile.

Urbackup DB is urbackup server with local clients (directory of backup, internet auth key and differents settings)
Mysql DB is mysql Database named Urbackup with table client_state who contain state of urbackup service on computer, internet auth key and client ID.

(cherry picked from commit de0a0440673b8068bd3050c46e7a3b271cbb034a)