(DEPLOY) Fix key error issues "KeyError: 'login'"
When a deployment is sent and the machine count calculation exceeds the duration of an XML-RPC session, the end of the PHP script that sends the deployments is lost. The consequence is that the Has_login_command table does not contain an entry for the command identifier. The deployment receives an error when it tries to retrieve this data.

Traceback (most recent call last):
  File "C:\Python27\Lib\site-packages\pulse_xmpp_agent\agentxmpp.py", line 2254, in message
    dataerreur
  File "C:\Python27\Lib\site-packages\pulse_xmpp_agent\lib\utils.py", line 561, in call_plugin
    pluginaction.action(*args, **kwargs)
  File "C:\Python27\Lib\site-packages\pulse_xmpp_agent\pluginsmachine\plugin_applicationdeploymentjson.py", line 621, in
 action
    sessionid)
  File "C:\Python27\Lib\site-packages\pulse_xmpp_agent\pluginsmachine\plugin_applicationdeploymentjson.py", line 2467, i
n recuperefilecdn
    takeresource(datasend, objectxmpp, sessionid)
  File "C:\Python27\Lib\site-packages\pulse_xmpp_agent\pluginsmachine\plugin_applicationdeploymentjson.py", line 2107, i
n takeresource
    "data" : {'user': datasendl['data']['advanced']['login'],
KeyError: 'login'

(cherry picked from commit b7b6f7b92b7e8cdb4b86de8752c896f0377672fd)