Revert "(auth) disable check une configuration file more useful"
This reverts commit 575537f3b3aed1924367080868ffb697f73ca941.

This untested commit broke the provisionning.

In the authentication process, we check on the config files if the user
is on the authonly or exclude sections.

We add this process back and if we want to avoid modifying by hand we
will have to think of an other way to do.