(imaging) profiles and postinstalls were not reinitialized when we clicked on 'reset entity bootmenus'
This was due to a confusion between reset all menus function and delete computer from imaging server.
In reset all menus, the menus are not completly destroyed.
When we delete a computer from imaging the Target and Menu are destroyed. If we subscribe again the machine it will copy the default menu out of the blue.